minor cast changes

+++ generic/tclIndexObj.c
@@ -183,7 +183,7 @@ Tcl_GetIndexFromObjStruct(interp, objPtr, tablePtr, offset, msg, flags,
     }
     
     for (entryPtr = tablePtr, i = 0; *entryPtr != NULL; 
-	    entryPtr = (char **) ((long) entryPtr + offset), i++) {
+	    entryPtr = (char **) ((size_t) entryPtr + offset), i++) {
 	for (p1 = key, p2 = *entryPtr; *p1 == *p2; p1++, p2++) {
 	    if (*p1 == 0) {
 		index = i;
@@ -229,10 +229,10 @@ Tcl_GetIndexFromObjStruct(interp, objPtr, tablePtr, offset, msg, flags,
 	Tcl_AppendStringsToObj(resultPtr,
 		(numAbbrev > 1) ? "ambiguous " : "bad ", msg, " \"",
 		key, "\": must be ", *tablePtr, (char *) NULL);
-	for (entryPtr = (char **) ((long) tablePtr + offset), count = 0;
+	for (entryPtr = (char **) ((size_t) tablePtr + offset), count = 0;
 		*entryPtr != NULL;
-		entryPtr = (char **) ((long) entryPtr + offset), count++) {
-	    if ((*((char **) ((long) entryPtr + offset))) == NULL) {
+		entryPtr = (char **) ((size_t) entryPtr + offset), count++) {
+	    if ((*((char **) ((size_t) entryPtr + offset))) == NULL) {
 		Tcl_AppendStringsToObj(resultPtr,
 			(count > 0) ? ", or " : " or ", *entryPtr,
 			(char *) NULL);
